-t option to zero out timestamps for reproducible builds
Most Android APK builds are quite close to being exactly reproducible, one
key failure there is the changing timestamps.  This adds a -t option which
resets all timestamps in the ZIP headers to in effect zero, but a slightly
different number to be safe with all timezones vs. the earliest date that
ZIP can represent.

This could easily be extended to support the standardized SOURCE_DATE_EPOCH
environment variable.  That would mean that zipalign seens that env var, it
would instead use that timestamp in the ZIP header.

https://reproducible-builds.org/

Change-Id: If4ea705626955378fa71caaeae7ca8318fac02fc